+++ common/src/main/java/house/greenhouse/bovinesandbuttercups/content/data/edible/BovinesEdibleBlockTypes.java
@@ -2,12 +2,8 @@ package house.greenhouse.bovinesandbuttercups.content.data.edible;
 
 import house.greenhouse.bovinesandbuttercups.BovinesAndButtercups;
 import house.greenhouse.bovinesandbuttercups.api.block.EdibleBlockType;
-import house.greenhouse.bovinesandbuttercups.content.data.nectar.BovinesNectars;
-import house.greenhouse.bovinesandbuttercups.content.data.nectar.Nectar;
 import house.greenhouse.bovinesandbuttercups.content.data.nectar.NectarEffects;
 import house.greenhouse.bovinesandbuttercups.registry.BovinesRegistryKeys;
-import net.minecraft.core.Holder;
-import net.minecraft.core.registries.Registries;
 import net.minecraft.data.worldgen.BootstrapContext;
 import net.minecraft.resources.ResourceKey;
 import net.minecraft.world.effect.MobEffects;
@@ -17,6 +13,7 @@ import java.util.List;
 public class BovinesEdibleBlockTypes {
     public static final ResourceKey<EdibleBlockType> BIRD_OF_PARADISE_CUPCAKE = ResourceKey.create(BovinesRegistryKeys.EDIBLE_BLOCK_TYPE, BovinesAndButtercups.asResource("bird_of_paradise_cupcake"));
     public static final ResourceKey<EdibleBlockType> BUTTERCUP_CUPCAKE = ResourceKey.create(BovinesRegistryKeys.EDIBLE_BLOCK_TYPE, BovinesAndButtercups.asResource("buttercup_cupcake"));
+    public static final ResourceKey<EdibleBlockType> CAMELLIA_CUPCAKE = ResourceKey.create(BovinesRegistryKeys.EDIBLE_BLOCK_TYPE, BovinesAndButtercups.asResource("camellia_cupcake"));
     public static final ResourceKey<EdibleBlockType> CHARGELILY_CUPCAKE = ResourceKey.create(BovinesRegistryKeys.EDIBLE_BLOCK_TYPE, BovinesAndButtercups.asResource("chargelily_cupcake"));
     public static final ResourceKey<EdibleBlockType> FREESIA_CUPCAKE = ResourceKey.create(BovinesRegistryKeys.EDIBLE_BLOCK_TYPE, BovinesAndButtercups.asResource("freesia_cupcake"));
     public static final ResourceKey<EdibleBlockType> HYACINTH_CUPCAKE = ResourceKey.create(BovinesRegistryKeys.EDIBLE_BLOCK_TYPE, BovinesAndButtercups.asResource("hyacinth_cupcake"));
@@ -34,6 +31,7 @@ public class BovinesEdibleBlockTypes {
     public static void bootstrap(BootstrapContext<EdibleBlockType> context) {
         context.register(BIRD_OF_PARADISE_CUPCAKE, EdibleBlockType.cupcake(context, new NectarEffects(List.of(new NectarEffects.Entry(MobEffects.SLOW_FALLING, 7200)))));
         context.register(BUTTERCUP_CUPCAKE, EdibleBlockType.cupcake(context, new NectarEffects(List.of(new NectarEffects.Entry(MobEffects.POISON, 2400)))));
+        context.register(CAMELLIA_CUPCAKE, EdibleBlockType.cupcake(context, new NectarEffects(List.of(new NectarEffects.Entry(MobEffects.HUNGER, 2400)))));
         context.register(CHARGELILY_CUPCAKE, EdibleBlockType.cupcake(context, new NectarEffects(List.of(new NectarEffects.Entry(MobEffects.DIG_SPEED, 1200)))));
         context.register(FREESIA_CUPCAKE, EdibleBlockType.cupcake(context, new NectarEffects(List.of(new NectarEffects.Entry(MobEffects.WATER_BREATHING, 9600)))));
         context.register(HYACINTH_CUPCAKE, EdibleBlockType.cupcake(context, new NectarEffects(List.of(new NectarEffects.Entry(MobEffects.WITHER, 2400)))));