+++ CHANGELOG.md
@@ -1,2 +1,4 @@
-## Bugfixes
-- [NEOFORGE] Fixed startup crash.
+## Changes
+- Added Camellia Mooblooms, a new magenta Moobloom bred whilst near Cherry Grove blocks.
+  - Pink Petals and Cherry Wood has been removed from the Pink Daisy breeding table to account for this change.
+- The Full Bloom advancement now grants 100 experience upon completion.